+++ mod/Patches/LeaderboardController.cs
@@ -15,4 +15,18 @@ namespace ArchipelagoULTRAKILL.Patches
             else return true;
         }
     }
+
+    [HarmonyPatch(typeof(LeaderboardController), "SubmitCyberGrindScore")]
+    class LeaderboardController_SubmitCyberGrindScore_Patch
+    {
+        public static bool Prefix()
+        {
+            if (Core.DataExists())
+            {
+                Core.logger.LogInfo("Current save file is randomized. Skipped leaderboard submission.");
+                return false;
+            }
+            else return true;
+        }
+    }
 }